Merge: Merged Branch Main into branch feat/homepage, making adjustments on function clear() in order to properly run tests
